+++ ExperimentalEnemyInteractions/EnemyPatches/NutCrackerAIPatch.cs
@@ -64,6 +64,7 @@ namespace NaturalSelection.EnemyPatches
         static void NutcrackerUpdatePostfix(NutcrackerEnemyAI __instance)
         {
             if (__instance.isEnemyDead) return;
+            CheckDataIntegrityNutcracker(__instance);
             NutcrackerData data = NutcrackerData[__instance];
 
             enemyList = LibraryCalls.GetCompleteList(__instance);
@@ -141,6 +142,7 @@ namespace NaturalSelection.EnemyPatches
         [HarmonyPostfix]
         static void DoAIIntervalPatch(NutcrackerEnemyAI __instance)
         {
+            CheckDataIntegrityNutcracker(__instance);
             NutcrackerData data = NutcrackerData[__instance];
 
             if (__instance.currentBehaviourStateIndex == 2)
@@ -164,6 +166,15 @@ namespace NaturalSelection.EnemyPatches
                         __instance.SwitchToBehaviourState(1);
                     }
                 }
+            }
+        }
+
+        public static void CheckDataIntegrityNutcracker(NutcrackerEnemyAI __instance)
+        {
+            if (!NutcrackerData.ContainsKey(__instance))
+            {
+                Script.Logger.Log(LogLevel.Fatal, $"Critical failule. Failed to get data for {LibraryCalls.DebugStringHead(__instance)}. Attempting to fix...");
+                NutcrackerData.Add(__instance, new NutcrackerData());
             }
         }
     }