+++ ExperimentalEnemyInteractions/EnemyPatches/PufferAIPatch.cs
@@ -47,6 +47,7 @@ namespace NaturalSelection.EnemyPatches
         static bool PrefixAIInterval(PufferAI __instance)
         {
             if (__instance.isEnemyDead) return true;
+            CheckDataIntegrityPuffer(__instance);
             PufferData pufferData = pufferList[__instance];
 
             if (__instance.currentBehaviourStateIndex == 2 && pufferData.targetEnemy != null && (Vector3.Distance(__instance.closestSeenPlayer.transform.position, __instance.transform.position) < Vector3.Distance(pufferData.targetEnemy.transform.position, __instance.transform.position)))
@@ -66,6 +67,7 @@ namespace NaturalSelection.EnemyPatches
         static void PostfixAIInterval(PufferAI __instance)
         {
             if (__instance.isEnemyDead) return;
+            CheckDataIntegrityPuffer(__instance);
             PufferData pufferData = pufferList[__instance];
 
             if (__instance.currentBehaviourStateIndex == 2 && pufferData.targetEnemy != null && (Vector3.Distance(__instance.closestSeenPlayer.transform.position, __instance.transform.position) < Vector3.Distance(pufferData.targetEnemy.transform.position, __instance.transform.position)))
@@ -90,6 +92,15 @@ namespace NaturalSelection.EnemyPatches
                 {
                     instance.KillEnemy(true);
                 }
+        }
+
+        public static void CheckDataIntegrityPuffer(PufferAI __instance)
+        {
+            if (!pufferList.ContainsKey(__instance))
+            {
+                Script.Logger.Log(LogLevel.Fatal, $"Critical failule. Failed to get data for {LibraryCalls.DebugStringHead(__instance)}. Attempting to fix...");
+                pufferList.Add(__instance, new PufferData());
+            }
         }
 
         public static void HitEnemyTest(int force, EnemyAI enemyWhoHit, bool playHitSFX, PufferAI instance)